+++ R/render.R
@@ -121,8 +121,7 @@ moon_reader = function(
 
     function(x, options) {
       res = hook_source(x, options)
-      # replace {{code}} with *code so that this line can be highlighted
-      gsub('(^|\n)([ \t]*)\\{\\{([^\n]+?)\\}\\}', '\\1\\2*\\3', res)
+      highlight_code(res)
     }
   }