+++ R/utils.R
@@ -121,3 +121,17 @@ summon_remark = function(version = 'latest', to = 'libs/') {
     file.path(to, name)
   )
 }
+
+# replace {{code}} with *code so that this line can be highlighted in remark.js;
+# this also works with multiple lines
+highlight_code = function(x) {
+  x = paste0('\n', x)  # prepend \n and remove it later
+  r = '(\n)([ \t]*)\\{\\{(.+?)\\}\\}'
+  m = gregexpr(r, x)
+  regmatches(x, m) = lapply(regmatches(x, m), function(z) {
+    z = gsub(r, '\\1\\2\\3', z)  # remove {{ and }}
+    z = gsub('\n', '\n*', z)     # add * after every \n
+    z
+  })
+  gsub('^\n', '', x)
+}