+++ NEWS.md
@@ -2,6 +2,8 @@
 
 ## BUG FIXES
 
+- Line highlighting using `{{}}` does not work with multiple lines (thanks, @HeidiSeibold #53 and @aj2duncan #54).
+
 - The option `mathjax: null` does not work for `moon_reader()`, i.e., it was not possible to exclude MathJax.
 
 # CHANGES IN xaringan VERSION 0.3